% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Title}
\usage{
check_data(.s, .t)
}
\arguments{
\item{.s}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.t}{The Traget Dataframe.
Must contain a unique column id and the columns you want to match on}
}
\value{
Messages
}
\description{
Description
}
\examples{
check_data(source, target)
}
