% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_data.R
\name{dedup_data}
\alias{dedup_data}
\title{Deduplicate Data}
\usage{
dedup_data(.score, .source, .target, .col)
}
\arguments{
\item{.score}{Dataframe generated by scores_data()}

\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Traget Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.col}{score column generated by scores_data()}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
tab_source <- table_source[1:100, ]
tab_target <- table_target[1:999, ]
tab_match <- match_data(
  .source = table_source[1:100, ],
  .target = table_target[1:999, ],
  .cols = c("name", "iso3", "city"),
  .min_sim = .2,
  .max_match = 10,
  .method = "osa", 
  .progress = TRUE
)
tab_score <- scores_data(tab_match)
dedup_data(tab_score, tab_source, tab_target, "score_mean")
}
