% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_legal_form.R
\name{extract_legal_form}
\alias{extract_legal_form}
\title{Extract Legal Forms}
\usage{
extract_legal_form(
  .tab,
  .col_name,
  .col_country = NULL,
  .legal_forms = data.frame(),
  .workers = future::availableCores()
)
}
\arguments{
\item{.tab}{A dataframe (either the source or target dataframe)}

\item{.col_name}{The column with firm names}

\item{.col_country}{Optionally, a column with iso3 country codes}

\item{.legal_forms}{A dataframe with legal forms}

\item{.workers}{Number of cores to utilize (Default all cores determined by future::availableCores())}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
extract_legal_form(
  .tab = table_source[1:100, ], 
  .col_name = "name", 
  .col_country = "iso3",
  .workers = 1
  )
}
