% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicates.R
\name{filter_duplicates}
\alias{filter_duplicates}
\title{Title}
\usage{
filter_duplicates(.data, ...)
}
\arguments{
\item{.data}{A dataframe}

\item{...}{Any number of columns to check}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
# filter_duplicates()
}
