% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights}
\alias{get_weights}
\title{Get Weights}
\usage{
get_weights(.tab, .cols_match)
}
\arguments{
\item{.tab}{A dataframe (either the source or target dataframe)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}
}
\value{
A numeric vector
}
\description{
Description
}
\examples{
get_weights(table_source, c("name", "city", "address"))
}
