% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{join_data}
\alias{join_data}
\title{Perform LeftJoin on Data}
\usage{
join_data(.source, .target, .cols, .join, .method = "osa")
}
\arguments{
\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Target Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.cols}{The column names to match as character vector}

\item{.join}{Columns to match on}

\item{.method}{One of "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"}
}
\value{
A Dataframe
}
\description{
Description
}
\examples{
join_data(
  .source = table_source,
  .target = table_target,
  .cols = c("name", "iso3", "city", "address"),
  .join = c("name", "iso3"),
  .method = "osa"
)
}
