% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-na.R
\docType{data}
\name{legal_form_ecb}
\alias{legal_form_ecb}
\title{Dataset: Legal Forms from the ECB}
\format{
A data frame with 2,399 rows and 3 variables:
\describe{
\item{iso3}{ISO 3166-1 alpha-3 country codes}
\item{legal_form_orig}{Original legal form}
\item{legal_form_stand}{Standardized legal form}
}
}
\usage{
legal_form_ecb
}
\description{
Dataset: Legal Forms from the ECB
}
\keyword{datasets}
