% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-na.R, R/datasets.R
\docType{data}
\name{legal_form_gleif}
\alias{legal_form_gleif}
\title{Dataset: Legal Forms from the GLEIF}
\format{
A data frame with 3,322 rows and 3 variables:
\describe{
\item{iso3}{ISO 3166-1 alpha-3 country codes}
\item{legal_form_orig}{Original legal form}
\item{legal_form_stand}{Standardized legal form}
}

A data frame with 3,322 rows and 3 variables:
\describe{
\item{iso3}{ISO 3166-1 alpha-3 country codes}
\item{legal_form_orig}{Original legal form}
\item{legal_form_stand}{Standardized legal form}
}
}
\usage{
legal_form_gleif

legal_form_gleif
}
\description{
Dataset: Legal Forms from the GLEIF

Dataset: Legal Forms from the GLEIF
}
\keyword{datasets}
