% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique.R
\name{make_unique}
\alias{make_unique}
\title{Make Dataframe Unique}
\usage{
make_unique(.data, ..., .rem = c("all", "last"))
}
\arguments{
\item{.data}{A dataframe}

\item{...}{Columns on which the dataframe should be unique}

\item{.rem}{"all": All duplicates are removed (also the first non-duplicated entry)\cr
"last": Only subsequent duplicates are removed (first value remains)}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
make_unique(table_source, name)
}
