% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_complete.R
\name{match_complete}
\alias{match_complete}
\title{Complete Match}
\usage{
match_complete(
  .source,
  .target,
  .cols_match,
  .cols_join = NULL,
  .cols_exact = NULL,
  .max_match = 10,
  .method = "osa",
  .verbose = TRUE,
  .workers = future::availableCores(),
  .char_block = c(Inf, Inf),
  .standardize = TRUE,
  .w_unique = NULL,
  .w_custom = NULL,
  .col_score = c("sms", "smw", "smc", "sss", "ssw", "ssc")
)
}
\arguments{
\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}

\item{.cols_join}{Columns to perfrom an exact match on, before fuzzy-matching.\cr
(Matched IDs will be excluded from fuzzy-match)}

\item{.cols_exact}{Columns that must be matched perfectly.\cr
(Data will be partitioned using those columns)}

\item{.max_match}{Maximum number of matches to return (Default = 10)}

\item{.method}{One of "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex".\cr
See: stringdist-metrics {stringdist}}

\item{.verbose}{Print additional information?}

\item{.workers}{Number of cores to utilize (Default all cores determined by future::availableCores())}

\item{.char_block}{Character Block Size. Used to partition data.\cr
\itemize{
\item First element chunks the source data in ngram-blocks.\cr
\item Second element allows for characters in target below/above block size.
}}

\item{.standardize}{Perform String Standardization using standardize_data()?}

\item{.w_unique}{Weights calculated by get_weights()}

\item{.w_custom}{A named numeric vector that matches the columns of .cols_match w/o the columns of .cols_exact}

\item{.col_score}{Score column generated by scores_data().\cr
Options are:\cr
sms: Simple Mean (mean over all fuzzy columns)\cr
smw: Weighted Mean (mean over all fuzzy columns, weighted by get_weights())\cr
smc: Custom Mean (mean over all fuzzy columns, weighted custom weights)\cr
sss: Simple Mean, squared (mean over all fuzzy columns, scores are squared)\cr
ssw: Weighted Mean, squared (mean over all fuzzy columns, scores are squared before weighted by get_weights())\cr
ssc: Custom Mean, squared (mean over all fuzzy columns, scores are squared before weighted custom weights)}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
match_complete(
  .source = table_source[1:100, ],
  .target = table_target[1:999, ],
  .cols_match = c("name", "iso3", "city", "address"),
  .cols_join = c("name", "iso3"),
  .cols_exact = "iso3",
  .max_match = 25,
  .method = "soundex",
  .verbose = TRUE,
  .workers = 4,
  .char_block = c(5, 5),
  .standardize = TRUE,
  .w_unique = NULL,
  .w_custom = c(name = .7, city = .2, address = .1),
  .col_score = "sms"
)
}
