% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data.R
\name{match_data}
\alias{match_data}
\title{Match Data}
\usage{
match_data(
  .source,
  .target,
  .cols,
  .join = NULL,
  .must_match = NULL,
  .max_match = 10,
  .min_sim = 0.2,
  .method = "osa",
  .chunk = 1,
  .progress = TRUE
)
}
\arguments{
\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Target Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.cols}{The column names to match as character vector}

\item{.join}{Columns to match on}

\item{.must_match}{Columns that must be matched perfectly}

\item{.max_match}{Maximum number of matches to return (Default = 10)}

\item{.min_sim}{Minimum Similarity as defined by the chosen method}

\item{.method}{One of "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"}

\item{.chunk}{Chunk your data}

\item{.progress}{Should a progress bar be shown?
See stringdist package}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
match_data(
  .source = table_source[1:100, ],
  .target = table_target[1:999, ],
  .cols = c("name", "iso3", "city"),
  .min_sim = .2,
  .max_match = 10,
  .must_match = "iso3",
  .method = "osa", 
  .progress = TRUE
)
}
