% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmatch-package.R
\docType{package}
\name{Rmatch-package}
\alias{Rmatch}
\alias{Rmatch-package}
\title{Rmatch: A General Package For Any Matching Tasks}
\description{
Fuzzy match two dataframes
}
\author{
\strong{Maintainer}: Matthias Uckert \email{Matthias.Uckert@Outlook.com}

}
\keyword{internal}
