% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_data.R
\name{scores_data}
\alias{scores_data}
\title{Score Data}
\usage{
scores_data(
  .matches,
  .source,
  .target,
  .cols_match,
  .cols_exact = NULL,
  .w_unique = NULL,
  .w_custom = NULL
)
}
\arguments{
\item{.matches}{Dataframe produced by match_data()}

\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}

\item{.cols_exact}{Columns that must be matched perfectly.\cr
(Data will be partitioned using those columns)}

\item{.w_unique}{Weights calculated by get_weights()}

\item{.w_custom}{A named numeric vector that matches the columns of .cols_match w/o the columns of .cols_exact}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
tab_source <- table_source[1:100, ]
tab_target <- table_target[1:999, ]
cols_match <- c("name", "iso3", "city", "address")
cols_exact <- "iso3"
cols_join  <- c("name", "iso3")
tab_match <- match_data(
  .source = tab_source,
  .target = tab_target,
  .cols_match = cols_match,
  .cols_exact = cols_exact,
  .cols_join = cols_join,
  .method = "soundex",
)
scores_data(
  .matches = tab_match, 
  .source = tab_source, 
  .target = tab_target, 
  .cols_match = cols_match,
  .cols_exact = cols_exact
  )
}
