% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_data.R
\name{standardize_data}
\alias{standardize_data}
\title{Standardize Data}
\usage{
standardize_data(.tab, .cols_match, .fun = NULL)
}
\arguments{
\item{.tab}{A dataframe (either the source or target dataframe)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}

\item{.fun}{Function for standardization, if NULL standardize_str() is used}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
standardize_data(table_source, c("name", "iso3", "city", "address"))
}
