% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-na.R, R/datasets.R
\docType{data}
\name{table_matches}
\alias{table_matches}
\title{Example Dataset: Matches}
\format{
A data frame with 2,426 rows and 11 variables:
\describe{
\item{id_s}{Source Dataframe: A unique identifier}
\item{id_t}{Traget Dataframe: A unique identifier}
\item{name_s}{Source Dataframe: Company name}
\item{name_t}{Traget Dataframe: Company name}
\item{iso3_s}{Source Dataframe: ISO 3166-1 alpha-3 country codes}
\item{iso3_t}{Traget Dataframe: ISO 3166-1 alpha-3 country codes}
\item{city_s}{Source Dataframe: City Name}
\item{city_t}{Source Dataframe: City Name}
\item{address_s}{Source Dataframe: Address Name}
\item{address_t}{Traget Dataframe: Address Name}
\item{match}{Indicator for matched names (always 1)}
}

A data frame with 2,426 rows and 11 variables:
\describe{
\item{id_s}{Source Dataframe: A unique identifier}
\item{id_t}{Traget Dataframe: A unique identifier}
\item{name_s}{Source Dataframe: Company name}
\item{name_t}{Traget Dataframe: Company name}
\item{iso3_s}{Source Dataframe: ISO 3166-1 alpha-3 country codes}
\item{iso3_t}{Traget Dataframe: ISO 3166-1 alpha-3 country codes}
\item{city_s}{Source Dataframe: City Name}
\item{city_t}{Source Dataframe: City Name}
\item{address_s}{Source Dataframe: Address Name}
\item{address_t}{Traget Dataframe: Address Name}
\item{match}{Indicator for matched names (always 1)}
}
}
\usage{
table_matches

table_matches
}
\description{
Example Dataset: Matches

Example Dataset: Matches
}
\keyword{datasets}
