% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-na.R, R/datasets.R
\docType{data}
\name{table_source}
\alias{table_source}
\title{Example Dataset: Source}
\format{
A data frame with 2,426 rows and 5 variables:
\describe{
\item{id}{A unique identifier}
\item{name}{Company name}
\item{iso3}{ISO 3166-1 alpha-3 country codes}
\item{city}{City Name}
\item{address}{Address Name}
}

A data frame with 2,426 rows and 5 variables:
\describe{
\item{id}{A unique identifier}
\item{name}{Company name}
\item{iso3}{ISO 3166-1 alpha-3 country codes}
\item{city}{City Name}
\item{address}{Address Name}
}
}
\usage{
table_source

table_source
}
\description{
Example Dataset: Source

Example Dataset: Source
}
\keyword{datasets}
