% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-na.R, R/datasets.R
\docType{data}
\name{table_target}
\alias{table_target}
\title{Example Dataset: Target}
\format{
A data frame with 4,671 rows and 5 variables:
\describe{
\item{id}{A unique identifier}
\item{name}{Company name}
\item{iso3}{ISO 3166-1 alpha-3 country codes}
\item{city}{City Name}
\item{address}{Address Name}
}

A data frame with 4,671 rows and 5 variables:
\describe{
\item{id}{A unique identifier}
\item{name}{Company name}
\item{iso3}{ISO 3166-1 alpha-3 country codes}
\item{city}{City Name}
\item{address}{Address Name}
}
}
\usage{
table_target

table_target
}
\description{
Example Dataset: Target

Example Dataset: Target
}
\keyword{datasets}
