% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-na.R
\docType{data}
\name{target}
\alias{target}
\title{Example Dataset: Target}
\format{
A data frame with 4,671 rows and 5 variables:
\describe{
\item{id}{A unique identifier}
\item{name}{Company name}
\item{iso3}{ISO 3166-1 alpha-3 country codes}
\item{city}{City Name}
\item{address}{Address Name}
}
}
\usage{
target
}
\description{
Example Dataset: Target
}
\keyword{datasets}
