% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness_data.R
\name{uniqueness_data}
\alias{uniqueness_data}
\title{Title}
\usage{
uniqueness_data(.data, .cols, .type = c("source", "target"))
}
\arguments{
\item{.data}{A Dataframe}

\item{.cols}{Column as character vector}

\item{.type}{either source or target}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
uniqueness_data(source, c("name", "iso3", "city", "address"), "source")
}
