% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check Source and Target Dataframe}
\usage{
check_data(.source, .target, .check = c("source", "all"), .html = TRUE)
}
\arguments{
\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Target Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.check}{Check only column that are also in source, or all columns}

\item{.html}{Output Check as hatml table?}
}
\value{
Messages
}
\description{
Description
}
\examples{
check_data(table_source, table_target)
}
