% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_data.R
\name{dedup_data}
\alias{dedup_data}
\title{Deduplicate Data}
\usage{
dedup_data(.score, .source, .target, .col = "mean_weight")
}
\arguments{
\item{.score}{Dataframe generated by scores_data()}

\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Target Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.col}{score column generated by scores_data()}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
tab_source <- table_source[1:100, ]
tab_target <- table_target[1:999, ]
cols <- c("name", "iso3", "city", "address")
must <- "iso3"
tab_match <- match_data(
  .source = tab_source,
  .target = tab_target,
  .cols = cols,
  .must_match = must,
)
tab_score <- scores_data(
  .matches = tab_match,
  .source = tab_source,
  .target = tab_target,
  .must_match = must
)

dedup_data(
  .score = tab_score, 
  .source = tab_source, 
  .target = tab_target, 
  )
}
