% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_data.R
\name{dedup_data}
\alias{dedup_data}
\title{Deduplicate Data}
\usage{
dedup_data(
  .score,
  .source,
  .target,
  .cols_match,
  .col_score = c("sms", "smw", "smc", "sss", "ssw", "ssc")
)
}
\arguments{
\item{.score}{Dataframe generated by scores_data()}

\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}

\item{.col_score}{Score column generated by scores_data().\cr
Options are:\cr
\itemize{
\item sms: Simple Mean (mean over all fuzzy columns)\cr
\item smw: Weighted Mean (mean over all fuzzy columns, weighted by get_weights())\cr
\item smc: Custom Mean (mean over all fuzzy columns, weighted custom weights)\cr
\item sss: Simple Mean, squared (mean over all fuzzy columns, scores are squared)\cr
\item ssw: Weighted Mean, squared (mean over all fuzzy columns, scores are squared before weighted by get_weights())\cr
\item ssc: Custom Mean, squared (mean over all fuzzy columns, scores are squared before weighted custom weights)
}}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
tab_source <- table_source[1:100, ]
tab_target <- table_target[1:999, ]
cols_match <- c("name", "iso3", "city", "address")
cols_exact <- "iso3"
cols_join  <- c("name", "iso3")
tab_match <- match_data(
  .source = tab_source,
  .target = tab_target,
  .cols_match = cols_match,
  .cols_exact = cols_exact,
  .cols_join = cols_join,
  .method = "soundex"
)
tab_score <- scores_data(
  .matches = tab_match, 
  .source = tab_source, 
  .target = tab_target, 
  .cols_match = cols_match,
  .cols_exact = cols_exact
  )

dedup_data(
  .score = tab_score, 
  .source = tab_source, 
  .target = tab_target,
  .cols_match = cols_match,
  .col_score = "sms"
  )
}
