% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_data.R
\name{dedup_data}
\alias{dedup_data}
\title{Deduplicate Data}
\usage{
dedup_data(.score, .source, .target, .col)
}
\arguments{
\item{.score}{Dataframe generated by scores_data()}

\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Traget Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.col}{score column generated by scores_data()}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
# dedup_data()
}
