% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_legal_form.R
\name{extract_legal_form}
\alias{extract_legal_form}
\title{Extract Legal Forms}
\usage{
extract_legal_form(
  .data,
  .col,
  .legal_forms = data.frame(),
  .workers = future::availableCores()
)
}
\arguments{
\item{.data}{A dataframe}

\item{.col}{The column with firm names}

\item{.legal_forms}{A dataframe with legal forms}

\item{.workers}{Number of cores}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
extract_legal_form(table_source[1:100, ], "name", .workers = 1)
}
