% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicates.R
\name{filter_duplicates}
\alias{filter_duplicates}
\title{Filter duplicate values}
\usage{
filter_duplicates(.data, ...)
}
\arguments{
\item{.data}{A dataframe}

\item{...}{Columns on which the dataframe should be unique}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
filter_duplicates(table_source, name)
}
