% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights}
\alias{get_weights}
\title{Get Weights}
\usage{
get_weights(.data, .cols)
}
\arguments{
\item{.data}{A Dataframe}

\item{.cols}{columns}
}
\value{
A numeric vector
}
\description{
Description
}
\examples{
get_weights(table_source, c("name", "city", "address"))
}
