% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{join_data}
\alias{join_data}
\title{Perform LeftJoin on Data}
\usage{
join_data(.source, .target, .cols_match, .cols_join, .method = "osa")
}
\arguments{
\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}

\item{.cols_join}{Columns to perfrom an exact match on, before fuzzy-matching.\cr
(Matched IDs will be excluded from fuzzy-match)}

\item{.method}{One of "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex".\cr
See: stringdist-metrics {stringdist}}
}
\value{
A Dataframe
}
\description{
Description
}
\examples{
join_data(
  .source = table_source,
  .target = table_target,
  .cols_match = c("name", "iso3", "city", "address"),
  .cols_join = c("name", "iso3"),
  .method = "osa"
)
}
