% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique.R
\name{make_unique}
\alias{make_unique}
\title{Title}
\usage{
make_unique(.data, ..., .rem = c("all", "last"))
}
\arguments{
\item{.data}{A dataframe}

\item{...}{Any number of columns to check}

\item{.rem}{All or Last}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
# make_unique()
}
