% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data.R
\name{match_data}
\alias{match_data}
\title{Match Data}
\usage{
match_data(
  .source,
  .target,
  .cols_match,
  .cols_join = NULL,
  .cols_exact = NULL,
  .max_match = 10,
  .method = "osa",
  .verbose = TRUE,
  .workers = future::availableCores(),
  .char_block = c(Inf, Inf)
)
}
\arguments{
\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}

\item{.cols_join}{Columns to perfrom an exact match on, before fuzzy-matching.\cr
(Matched IDs will be excluded from fuzzy-match)}

\item{.cols_exact}{Columns that must be matched perfectly.\cr
(Data will be partitioned using those columns)}

\item{.max_match}{Maximum number of matches to return (Default = 10)}

\item{.method}{One of "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex".\cr
See: stringdist-metrics {stringdist}}

\item{.verbose}{Print additional information?}

\item{.workers}{Number of cores to utilize (Default all cores determined by future::availableCores())}

\item{.char_block}{Character Block Size. Used to partition data.\cr
\itemize{
\item First element chunks the source data in ngram-blocks.\cr
\item Second element allows for characters in target below/above block size.
}}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
tab_source <- table_source[1:100, ]
tab_target <- table_target[1:999, ]
cols_match <- c("name", "iso3", "city", "address")
cols_join  <- c("name", "iso3")
cols_exact <- "iso3"

match_data(
  .source = tab_source,
  .target = tab_target,
  .cols_match = cols_match,
  .cols_join = cols_join,
  .cols_exact = cols_exact
)
}
