% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data.R
\name{match_data}
\alias{match_data}
\title{Title}
\usage{
match_data(.s, .t, .cols, .max_match = 10, .min_sim = 0.8, .method = "osa")
}
\arguments{
\item{.s}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.t}{The Traget Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.cols}{The column names to match as character vector}

\item{.max_match}{Maximum number of matches to return (Default = 10)}

\item{.min_sim}{Minimum Similarity as defined by the chosen method}

\item{.method}{One of "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"
See stringdist package}
}
\value{
A Dataframe
}
\description{
Description
}
\examples{
library(Rmatch)
match_data(source[1:100, ], target[1:1000, ], c("name", "city"), .min_sim = 0)
}
