% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_data.R
\name{scores_data}
\alias{scores_data}
\title{Score Data}
\usage{
scores_data(
  .matches,
  .source,
  .target,
  .must_match = NULL,
  .w_unique = NULL,
  .w_custom = NULL
)
}
\arguments{
\item{.matches}{Dataframe produced by match_data()}

\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Target Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.must_match}{Columns that must be matched perfectly}

\item{.w_unique}{Weights calucalted by get_weights()}

\item{.w_custom}{Custom weights}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
tab_source <- table_source[1:100, ]
tab_target <- table_target[1:999, ]
cols <- c("name", "iso3", "city", "address")
must <- "iso3"
tab_match <- match_data(
  .source = tab_source,
  .target = tab_target,
  .cols = cols,
  .must_match = must,
)
scores_data(
  .matches = tab_match,
  .source = tab_source,
  .target = tab_target,
  .must_match = must
)
}
