% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_data.R
\name{scores_data}
\alias{scores_data}
\title{Score Data}
\usage{
scores_data(.data, .weights = NULL, .training = NULL)
}
\arguments{
\item{.data}{A dataframe}

\item{.weights}{Custom weights}

\item{.training}{A training dataset}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
tab_match <- match_data(
  .source = table_source[1:100, ],
  .target = table_target[1:999, ],
  .cols = c("name", "iso3", "city"),
  .min_sim = .2,
  .max_match = 10,
  .method = "osa", 
  .progress = TRUE
)
tab_score <- scores_data(tab_match)
}
