% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_data.R
\name{scores_data}
\alias{scores_data}
\title{Title}
\usage{
scores_data(.data, .cols_extact = c())
}
\arguments{
\item{.data}{A dataframe}

\item{.cols_extact}{Columns that need extact matching}
}
\value{
A dataframe
}
\description{
Description
}
