% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_data.R
\name{scores_data}
\alias{scores_data}
\title{Score Data}
\usage{
scores_data(.matches, .source, .target, .must_match = NULL, .weights = NULL)
}
\arguments{
\item{.matches}{Dataframe produced by match_data()}

\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Target Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.must_match}{Columns that must be matched perfectly}

\item{.weights}{Custom weights}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
tab_match <- match_data(
  .source = table_source,
  .target = table_target,
  .cols = c("name", "iso3", "city"),
  .min_sim = .2,
  .max_match = 10,
  .method = "osa", 
  .progress = TRUE
)
tab_score <- scores_data(tab_match, table_source, table_target)
}
