% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_data.R
\name{standardize_data}
\alias{standardize_data}
\title{Title}
\usage{
standardize_data(.data, .cols, .op = c("space", "punct", "case", "utf8"))
}
\arguments{
\item{.data}{A dataframe}

\item{.cols}{Columns to standardize}

\item{.op}{Operations to perform}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
standardize_data(source, c("name", "iso3", "city", "address"))
}
