% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_data.R
\name{standardize_data}
\alias{standardize_data}
\title{Standardize Data}
\usage{
standardize_data(.data, .cols, .fun = NULL)
}
\arguments{
\item{.data}{A dataframe}

\item{.cols}{Columns to standardize}

\item{.fun}{Function for standardization, if NULL standardize_str() is used}
}
\value{
A dataframe
}
\description{
Description
}
\examples{
standardize_data(table_source, c("name", "iso3", "city", "address"))
}
