% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_str.R
\name{standardize_str}
\alias{standardize_str}
\title{Standardize Strings}
\usage{
standardize_str(.str, .op = c("space", "punct", "case", "utf8"))
}
\arguments{
\item{.str}{A string}

\item{.op}{Operations to perform}
}
\value{
A string
}
\description{
Description
}
\examples{
standardize_str(c("jkldsa   jkdhas   äää  §$ ## #'''"))
}
